% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rba_decision_table.R
\name{read_rba_decision_table}
\alias{read_rba_decision_table}
\title{Read a table of RBA monetary policy decisions
https://www.rba.gov.au/statistics/cash-rate/}
\usage{
read_rba_decision_table()
}
\description{
Read a table of RBA monetary policy decisions
https://www.rba.gov.au/statistics/cash-rate/
}
\keyword{internal}
