% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rba_decisions.R
\name{read_rba_decisions}
\alias{read_rba_decisions}
\title{Scrape RBA monetary policy decision media releases in a tidy tibble}
\usage{
read_rba_decisions(refresh = TRUE)
}
\arguments{
\item{refresh}{If \code{TRUE} (the default), the RBA website will be scraped for
any decisions that are not included in the package's internal data. If
\code{FALSE}, only the, pre-scraped data will be returned.}
}
\description{
Scrape RBA monetary policy decision media releases in a tidy tibble
}
\examples{
\dontrun{
# Scrape all monetary policy decisions from the RBA website
all_decisions <- read_rba_decisions()
}
}
\author{
Matt Cowgill
}
