% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rba_minutes.R
\name{read_rba_minutes}
\alias{read_rba_minutes}
\title{Scrape minutes of RBA Board/Monetary Policy Board in a tidy tibble}
\usage{
read_rba_minutes(refresh = TRUE)
}
\arguments{
\item{refresh}{If \code{TRUE} (the default), the RBA website will be scraped for
any minutes that are not included in the package's internal data. If
\code{FALSE}, only the, pre-scraped data will be returned.}
}
\description{
Scrape minutes of RBA Board/Monetary Policy Board in a tidy tibble
}
\examples{
\dontrun{
# Scrape all monetary policy minutes from the RBA website
all_decisions <- read_rba_minutes()
}
}
\author{
Matt Cowgill
}
