% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rba_decisions.R
\name{read_rba_decisions}
\alias{read_rba_decisions}
\title{Scrape RBA monetary policy decision media releases in a tidy tibble}
\usage{
read_rba_decisions(refresh = TRUE)
}
\arguments{
\item{min_year}{If \code{NULL} (the default), all releases will be returned. If a
year is specified (eg. \code{2015}), only releases from that year onwards will be
scraped.}

\item{use_existing}{If \code{TRUE} (the default), pre-scraped data, stored as
the object \code{rba_decisions}, will be used in the returned object, and only
newer decisions scraped from the RBA. If \code{FALSE}, all decisions will be
scraped fresh.}
}
\description{
Scrape RBA monetary policy decision media releases in a tidy tibble
}
\examples{
\dontrun{
# Scrape all monetary policy decisions from the RBA website
all_decisions <- read_rba_decisions()
}
}
\author{
Matt Cowgill
}
